requirements.txt
requests 
python-dotenv 
streamlit
openai
pymupdf
google-generativeai